% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hurdat.R
\name{get_hurdat2}
\alias{get_hurdat2}
\title{Download HURDAT2 Best Track}
\usage{
get_hurdat2(basin = c("AL", "EP"), storm_id)
}
\arguments{
\item{storm_id}{Storm name in all caps, dash, year (EX: "HARVEY-2017")}

\item{path}{Path to download files into (if not provided,
files will be download into the \code{tempdir()})}
}
\value{
HURDAT2 best track for a given storm
}
\description{
Downloads most recent HURDAT2 file from NOAA
\href{https://www.nhc.noaa.gov/data/hurdat/}{HURDAT2 HTTPS Server}
}
\examples{
\dontrun{
# get an individual huricane's data by name
get_hurdat2(basin = "AL", storm_id = "HARVEY-2017")

# get an individual huricane's data by ATCF ID
get_hurdat2(storm_id = "AL092017")

# get all data for a basin
get_hurdat2(basin = "AL")
}
}
