% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nClimGrid.R
\name{get_nClimGrid}
\alias{get_nClimGrid}
\title{Download and Filter Weather Data}
\usage{
get_nClimGrid(start, end)
}
\arguments{
\item{start}{start date to get data for}

\item{end}{end date to get date for}
}
\value{
terra raster of climate data over the interval specified
}
\description{
Downloads NOAA \href{https://www.ncei.noaa.gov/products/land-based-station/nclimgrid-daily}{nClimGrid-Daily}
data for the specified interval.
}
\examples{
\dontrun{
harvey_eaglei <- get_eaglei(
  interval = interval(harvey_start, harvey_end + days(5)),
  path = "raw-data/eaglei"
)
}
}
